Don't obscure error status in "read"
If a `read` promise is rejected, the rejection reason is an `Error` object with the message `[object Object]`. It's not possible to get the reponse.status.

This removes the `Error` and throws the response directly. Q supports rejection reasons that are not instances of `Error`.

Include response as a property on Error